requirements.txt
# Django Core
Django==4.2.9
django-tenants==3.6.1
psycopg2-binary==2.9.9

# REST API
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1

# Authentication
django-allauth==0.60.1

# Environment & Configuration
python-dotenv==1.0.0
dj-database-url==2.1.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-dateutil==2.8.2
requests==2.31.0

# Cloud Provider SDKs
oci==2.119.1
boto3==1.34.34
